+++ tests/test_version.py
@@ -50,7 +50,8 @@ class TestVersionModule(unittest.TestCase):
             or cur.startswith("2.1.")
             or cur.startswith("2.2.")
             or cur.startswith("2.3.")
-            or cur.startswith("2.4."),
+            or cur.startswith("2.4.")
+            or cur.startswith("2.5."),
             cur,
         )
         if "rc" in cur.lower():